backend_old/flask-python-backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
flask_sqlalchemy
flask_cors
pymongo
pyomngo[srv]